# Performance: ajout d'index sur la date de naissance pour fluidifier le parcours de création d'utilisateur