## :thinking: Pourquoi ?

Cf [Sentry](https://itou.sentry.io/performance/summary/?project=6164438&query=http.method%3APOST&referrer=performance-transaction-summary&statsPeriod=7d&transaction=%2Fapply%2F%7Bcompany_pk%7D%2Fsender%2Fcreate_job_seeker%2F%7Bsession_uuid%7D%2F1&unselectedSeries=p100%28%29&unselectedSeries=avg%28%29)

À la première étape de création, on vérifie si un candidat avec les mêmes nom/prénom/date de naissance n'existe pas déjà.

D'après [Sentry](https://itou.sentry.io/insights/database/spans/span/21b4e349878436bb/?project=6164438&statsPeriod=7d), cette vérification prend en moyenne 492 ms mais cela peut monter à plus d'1 seconde.

En local, la création de l'index fait passer la requête de ~65ms à ~1ms donc on devrait voir rapidement une amélioration.

<!--
# Catégories changelog

 +--------------------------|--------------------------+
 | API                      | Notifications            |
 | Accessibilité            | Page d’accueil           |
 | Admin                    | PASS IAE                 |
 | Annexes financières      | Performances             |
 | Candidature              | Pilotage                 |
 | Connexion                | Profil salarié           |
 | Contrôle a posteriori    | Prescripteur             |
 | Demandes de prolongation | Recherche employeur      |
 | Demandeur d’emploi       | Recherche fiche de poste |
 | Employeur                | Recherche prescripteur   |
 | Fiche de poste           | Stabilité                |
 | Fiche entreprise         | Statistiques             |
 | Fiches salarié           | Tableau de bord          |
 | GEIQ                     | UX/UI                    |
 | Inscription              | Vie privée               |
 +--------------------------|--------------------------+

-->

## :cake: Comment ? <!-- optionnel -->

This will be useful in CreateJobSeekerStep1ForSenderView where we check if job seeker with the same names & birth date already exists.

In Sentry, it can be seen that:

    SELECT users_user.id, users_user.password, users_user.last_login, users_user.is_superuser,
      users_user.username, users_user.first_name, users_user.last_name, users_user.is_staff,
      users_user.is_active, users_user.date_joined, users_user.address_line_1, users_user.address_line_2,
      users_user.post_code, users_user.city, users_user.department, users_user.coords,
      users_user.geocoding_score, users_user.geocoding_updated_at, users_user.ban_api_resolved_address,
      users_user.insee_city_id, users_user.title, users_user.email, users_user.phone, users_user.kind,
      users_user.identity_provider, users_user.has_completed_welcoming_tour, users_user.created_by_id,
      users_user.external_data_source_history, users_user.last_checked_at, users_user.public_id,
      users_user.address_filled_at, users_user.first_login
    FROM users_user
    INNER JOIN users_jobseekerprofile ON (users_user.id = users_jobseekerprofile.user_id)
    WHERE (
      UPPER(UNACCENT(users_user.first_name)::text) = UPPER((UNACCENT(%s))) AND
        users_jobseekerprofile.birthdate = %s AND users_user.kind = %s AND UPPER(UNACCENT(
        users_user.last_name)::text) = UPPER((UNACCENT(%s)))
    )
    ORDER BY users_user.id ASC
    LIMIT 1

has an average duration of 492ms
cf https://itou.sentry.io/insights/database/spans/span/21b4e349878436bb/?project=6164438&statsPeriod=7d

Locally, creating the index improves the query speed from ~65ms to ~1ms

## :rotating_light: À vérifier

- [ ] Mettre à jour le CHANGELOG_breaking_changes.md ?

## :desert_island: Comment tester

> _Les instructions pour reproduire le problème, les profils de test, le parcours spécifique à utiliser, etc. Si vous disposez d'une recette jetable, mettre l'URL pour tester dans cette partie._

## :computer: Captures d'écran <!-- optionnel -->